requirements-vision.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
Pillow>=10.0.0
psutil>=5.9.0
httpx>=0.28.0
pydantic>=2.10.0
numpy>=1.26.0
